Reorganize component architecture
