+++ src/App/App.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
-import './App.css';
-import MoviesContainer from './MoviesContainer';
-import InfoPage from './InfoPage';
-import { getMovies, getSelectedTrailer } from './apiCalls';
 import { NavLink, Route } from 'react-router-dom';
+import './App.css';
+import MoviesContainer from '../MoviesContainer/MoviesContainer';
+import InfoPage from '../InfoPage/InfoPage';
+import { getMovies, getSelectedTrailer } from '../apiCalls';
 
 class App extends Component {
   constructor() {
